Rearrange services
